requirements.txt
Flask==2.3.2
Flask-SQLAlchemy==3.1.1
redis==5.0.1
rq==1.15.1
pycryptodome==3.18.0
PyMySQL==1.1.0
python-dotenv==1.0.1
pytest==8.1.1
PyJWT